requirements.txt
aiogram
python-dateutil